+++ examples/workflow-concurrency/CHANGELOG.md
@@ -1,5 +1,131 @@
 # Changelog
 
+## [1.16.71-beta.4](https://github.com/AIGNE-io/aigne-framework/compare/example-workflow-concurrency-v1.16.71-beta.3...example-workflow-concurrency-v1.16.71-beta.4) (2025-09-24)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @aigne/agent-library bumped to 1.21.46-beta.4
+    * @aigne/cli bumped to 1.48.4-beta.4
+    * @aigne/core bumped to 1.61.0-beta.3
+    * @aigne/openai bumped to 0.16.0-beta.4
+  * devDependencies
+    * @aigne/test-utils bumped to 0.5.53-beta.3
+
+## [1.16.71-beta.3](https://github.com/AIGNE-io/aigne-framework/compare/example-workflow-concurrency-v1.16.71-beta.2...example-workflow-concurrency-v1.16.71-beta.3) (2025-09-23)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @aigne/agent-library bumped to 1.21.46-beta.3
+    * @aigne/cli bumped to 1.48.4-beta.3
+    * @aigne/core bumped to 1.61.0-beta.2
+    * @aigne/openai bumped to 0.16.0-beta.3
+  * devDependencies
+    * @aigne/test-utils bumped to 0.5.53-beta.2
+
+## [1.16.71-beta.2](https://github.com/AIGNE-io/aigne-framework/compare/example-workflow-concurrency-v1.16.71-beta.1...example-workflow-concurrency-v1.16.71-beta.2) (2025-09-23)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @aigne/agent-library bumped to 1.21.46-beta.2
+    * @aigne/cli bumped to 1.48.4-beta.2
+    * @aigne/core bumped to 1.61.0-beta.1
+    * @aigne/openai bumped to 0.16.0-beta.2
+  * devDependencies
+    * @aigne/test-utils bumped to 0.5.53-beta.1
+
+## [1.16.71-beta.1](https://github.com/AIGNE-io/aigne-framework/compare/example-workflow-concurrency-v1.16.71-beta...example-workflow-concurrency-v1.16.71-beta.1) (2025-09-23)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @aigne/agent-library bumped to 1.21.46-beta.1
+    * @aigne/cli bumped to 1.48.4-beta.1
+    * @aigne/openai bumped to 0.16.0-beta.1
+
+## [1.16.71-beta](https://github.com/AIGNE-io/aigne-framework/compare/example-workflow-concurrency-v1.16.70...example-workflow-concurrency-v1.16.71-beta) (2025-09-22)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @aigne/agent-library bumped to 1.21.46-beta
+    * @aigne/cli bumped to 1.48.4-beta
+    * @aigne/core bumped to 1.61.0-beta
+    * @aigne/openai bumped to 0.16.0-beta
+  * devDependencies
+    * @aigne/test-utils bumped to 0.5.53-beta
+
+## [1.16.70](https://github.com/AIGNE-io/aigne-framework/compare/example-workflow-concurrency-v1.16.70-beta...example-workflow-concurrency-v1.16.70) (2025-09-22)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @aigne/cli bumped to 1.48.3
+
+## [1.16.70-beta](https://github.com/AIGNE-io/aigne-framework/compare/example-workflow-concurrency-v1.16.69...example-workflow-concurrency-v1.16.70-beta) (2025-09-22)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @aigne/cli bumped to 1.48.3-beta
+
+## [1.16.69](https://github.com/AIGNE-io/aigne-framework/compare/example-workflow-concurrency-v1.16.68...example-workflow-concurrency-v1.16.69) (2025-09-18)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @aigne/cli bumped to 1.48.2
+
+## [1.16.68](https://github.com/AIGNE-io/aigne-framework/compare/example-workflow-concurrency-v1.16.67...example-workflow-concurrency-v1.16.68) (2025-09-18)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @aigne/agent-library bumped to 1.21.45
+    * @aigne/cli bumped to 1.48.1
+    * @aigne/core bumped to 1.60.3
+    * @aigne/openai bumped to 0.15.4
+  * devDependencies
+    * @aigne/test-utils bumped to 0.5.52
+
+## [1.16.67](https://github.com/AIGNE-io/aigne-framework/compare/example-workflow-concurrency-v1.16.66...example-workflow-concurrency-v1.16.67) (2025-09-16)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @aigne/cli bumped to 1.48.0
+
+## [1.16.66](https://github.com/AIGNE-io/aigne-framework/compare/example-workflow-concurrency-v1.16.65...example-workflow-concurrency-v1.16.66) (2025-09-16)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @aigne/cli bumped to 1.47.1
+
 ## [1.16.65](https://github.com/AIGNE-io/aigne-framework/compare/example-workflow-concurrency-v1.16.64...example-workflow-concurrency-v1.16.65) (2025-09-14)